// printer-cloud-new/PrinterFlow/components/TaskTemplates/FilterTaskTemplates/types.tsx
import { taskTemplateStatus } from '../../../../services/printer-flow/types/taskTemplate';

export interface FilterButtonTaskTemplatesContainerProps {
  children: React.ReactNode;
  params: FilterTaskTemplateParams;
  setParams: React.Dispatch<React.SetStateAction<FilterTaskTemplateParams>>;
}

export interface FilterButtonTaskTemplatesProps {
  children: React.ReactNode;
  params: FilterTaskTemplateParams;
  setParams: React.Dispatch<React.SetStateAction<FilterTaskTemplateParams>>;
}

export interface FilterTaskTemplateParams {
  order: string;
  direction: string;
  q: string;
  page: number;
  perPage: number;
  status: taskTemplateStatus;
}